Refactoring and simplifying TopSearchBar
