+++ public/app/core/components/NavBar/NavBar.tsx
@@ -12,8 +12,6 @@ import { Icon, useTheme2, CustomScrollbar } from '@grafana/ui';
 import { getKioskMode } from 'app/core/navigation/kiosk';
 import { StoreState } from 'app/types';
 
-import { OrgSwitcher } from '../OrgSwitcher';
-
 import NavBarItem from './NavBarItem';
 import { NavBarItemIcon } from './NavBarItemIcon';
 import { NavBarItemWithoutMenu } from './NavBarItemWithoutMenu';
@@ -39,14 +37,9 @@ export const NavBar = React.memo(() => {
   const theme = useTheme2();
   const styles = getStyles(theme);
   const location = useLocation();
-  const [showSwitcherModal, setShowSwitcherModal] = useState(false);
   const [menuOpen, setMenuOpen] = useState(false);
   const [menuAnimationInProgress, setMenuAnimationInProgress] = useState(false);
   const [menuIdOpen, setMenuIdOpen] = useState<string | undefined>(undefined);
-
-  const toggleSwitcherModal = () => {
-    setShowSwitcherModal(!showSwitcherModal);
-  };
 
   // Here we need to hack in a "home" and "search" NavModelItem since this is constructed in the frontend
   const searchItem: NavModelItem = enrichWithInteractionTracking(
@@ -79,8 +72,7 @@ export const NavBar = React.memo(() => {
     .map((item) => enrichWithInteractionTracking(item, menuOpen));
   const configItems = enrichConfigItems(
     navTree.filter((item) => item.section === NavSection.Config),
-    location,
-    toggleSwitcherModal
+    location
   ).map((item) => enrichWithInteractionTracking(item, menuOpen));
 
   const activeItem = isSearchActive(location) ? searchItem : getActiveItem(navTree, location.pathname);
@@ -159,7 +151,6 @@ export const NavBar = React.memo(() => {
           </FocusScope>
         </NavBarContext.Provider>
       </nav>
-      {showSwitcherModal && <OrgSwitcher onDismiss={toggleSwitcherModal} />}
       {(menuOpen || menuAnimationInProgress) && (
         <div className={styles.menuWrapper}>
           <NavBarMenu
